Integrated various changes worked on at the Glasgow Science Center. Changes include change the CameraManipulators so they work with double for time instead of float.  Also added support for DataType to osg::StateAttribute and StateSet so that they can be set to either STATIC or DYNAMIC, this allows the optimizer to know whether that an attribute can be optimized or not.

+++ src/osgUtil/FlightManipulator.cpp
@@ -175,7 +175,7 @@ bool FlightManipulator::calcMovement()
     if (_ga_t0.get()==NULL || _ga_t1.get()==NULL) return false;
 
 
-    float dt = _ga_t0->time()-_ga_t1->time();
+    double dt = _ga_t0->time()-_ga_t1->time();
 
     if (dt<0.0f)
     {
